refactor(commands): extract install/add/remove/outdated/upgrade into install.rs (Issue #271)
Continues the PR-A10 (Issue #186) module-split pattern. Moves
add/remove/install/outdated/upgrade command logic out of
src/commands/mod.rs into a new src/commands/install.rs, shrinking
mod.rs from 4903 to 3798 lines.
