+++ docs/PLAN.md
@@ -240,7 +240,9 @@
   - Tests: `cargo build` / `cargo clippy --all-targets --all-features -- -D warnings` / `cargo fmt -- --check` / `cargo test --lib`（354 passed）/ `cargo test`（フルスイート、E2E含む全テストパス）。JSON/CLI の公開挙動・schema に変更なし。
   - Current: `pybun doctor` / `pybun gc` の実装（`run_doctor` / `run_gc`、support bundle / PyPI cache stale check / PEP 723 cache GC 連携を含む）を `src/commands/maintenance.rs` へ抽出し、`mod.rs` 側の dispatcher を `maintenance::run_doctor(...)` / `maintenance::run_gc(...)` 呼び出しへ更新。`support_bundle` と cache size formatting の局所 import を新モジュール側へ移動し、JSON detail / diagnostics / summary は変更なし。`mod.rs` は 5241行 → 4903行に縮小。
   - Tests: `cargo check` / `cargo fmt -- --check` / `just lint` / `cargo test --test gc`（10 passed）/ `cargo test --test support_bundle`（2 passed）/ `cargo test --test self_update doctor`（4 passed, 10 filtered）/ `cargo test`（フルスイートパス）/ `cargo audit` で `gc` / `doctor` / support bundle 経路が継続動作することを確認。
-  - Remaining: `install.rs` / `lock.rs` / `run.rs` / `build.rs` / `x.rs` / `python_env.rs` / `project.rs`（init/outdated/upgrade）への抽出は follow-up PR で対応する。
+  - Current（Issue #271 follow-up）: `pybun add` / `pybun remove` / `pybun install` / `pybun outdated` / `pybun upgrade` の実装を `src/commands/install.rs` へ抽出し、`mod install;` を宣言。`mod.rs` 側の dispatcher を `install::run_add(...)` / `install::run_remove(...)` / `install::run_install(...)` / `install::run_outdated(...)` / `install::run_upgrade(...)` 呼び出しへ更新。新モジュールから直接参照される構造体・フィールドは `pub(crate)` 化（`super::` 経由の private アクセスでは不足したもののみ）。JSON detail / diagnostics / summary は変更なし。`mod.rs` は 4903行 → 3798行に縮小、新設の `install.rs` は 1662行。
+  - Tests: `cargo test --test '*'`（`cli_run` の一部失敗はローカル環境の壊れた Homebrew `python@3.14` に起因する既存の環境依存問題であり、`git stash` で本変更を除いた HEAD でも同一の失敗が再現することを確認済みのため本PRのスコープ外）/ `cargo test --test e2e_general`（9 passed）/ `just lint`（clippy + fmt check、warning なし）/ `cargo fmt -- --check`（クリーン）。
+  - Remaining: `lock.rs` / `run.rs` / `build.rs` / `x.rs` / `python_env.rs` / `project.rs`（init/outdated/upgrade）への抽出は follow-up PR で対応する。
 - [DONE] PR-A11: production `unwrap()` / `expect()` panic reduction pass (Issue #187)
   - Goal: 非テストの本番経路に残る panic を減らし、recoverable な状態を JSON diagnostics / Result 経路へ戻す。特に PyPI cache / verification / watcher / snapshot / module scan の周辺で、agent-facing な実行が unstructured panic で落ちないようにする。
   - Current: `src/pypi.rs` の stale cache notice Mutex が poison された場合でも `take_stale_cache_notices()` / stale notice push が panic せず `into_inner()` で復旧するよう変更し、`PyPiIndex::get()` の `cached.expect("checked")` を `Option` 分岐に置換。`src/commands/mod.rs::ensure_selection_is_verifiable()` の strict verification 後 hash 取得を `expect` から `E_VERIFY_MISSING_HASH` diagnostic 付き `Err` に変更。`src/commands/tooling.rs::run_watch()` の target unwrap、`src/module_finder.rs` の suffix unwrap、`src/snapshot.rs::load_for()` の HashMap get unwrap、`src/schema.rs::EventCollector::event()` の `last_mut().unwrap()`、`src/downloader.rs::Downloader::new()` の reqwest client build expect を panic しない実装へ置換。
